more py3k fixes. Localization analysis code (but probably not GUI) now works on python3

+++ PYME/localization/FitFactories/Interpolators/LinearInterpolator.py
@@ -112,19 +112,19 @@ class LinearInterpolator(__interpolator):
         #what region is 'safe' (ie we won't run out of model to interpret)
         #for these slices...
         if self.SplitPSF:
-            xm = len(X)/2
-            dx = min((interpolator.shape[1] - len(X))/2, xm) - 2
+            xm = int(len(X)/2)
+            dx = min(int((interpolator.shape[1] - len(X))/2), xm) - 2
     
-            ym = len(Y)/2
-            dy = min((interpolator.shape[1] - len(Y))/2, ym) - 2
+            ym = int(len(Y)/2)
+            dy = min(int((interpolator.shape[1] - len(Y))/2), ym) - 2
     
             safeRegion = ((X[xm-dx], X[xm+dx]), (Y[ym-dy], Y[ym+dy]),(Z[0] + self.IntZVals[2], Z[0] + self.IntZVals[-2]))
         else:
-            xm = len(X)/2
-            dx = min((interpolator.shape[0] - len(X))/2, xm) - 2
+            xm = int(len(X)/2)
+            dx = min(int((interpolator.shape[0] - len(X))/2), xm) - 2
     
-            ym = len(Y)/2
-            dy = min((interpolator.shape[1] - len(Y))/2, ym) - 2
+            ym = int(len(Y)/2)
+            dy = min(int((interpolator.shape[1] - len(Y))/2), ym) - 2
     
             safeRegion = ((X[xm-dx], X[xm+dx]), (Y[ym-dy], Y[ym+dy]),(Z[0] + self.IntZVals[2], Z[0] + self.IntZVals[-2]))
